+++ backend/hqlib/metric/project/project_management_metrics.py
@@ -29,10 +29,12 @@ class ActivityMetric(LowerIsBetterMetric):
     unit = 'dagen'
 
     def value(self) -> MetricValue:
-        return -1 if self._missing() else (datetime.datetime.now() - self._metric_source.datetime()).days
+        return -1 if self._missing() else \
+            (datetime.datetime.now() - self._metric_source.datetime(*self._get_metric_source_ids())).days
 
     def _missing(self) -> bool:
-        return self._metric_source.datetime() in (None, datetime.datetime.min)
+        return self._metric_source.datetime(*self._get_metric_source_ids()) in (None, datetime.datetime.min) \
+            if self._metric_source else True
 
 
 class RiskLog(ActivityMetric):
@@ -73,11 +75,13 @@ class OverDueActions(LowerIsBetterMetric):
     metric_source_class = metric_source.ActionLog
 
     def value(self) -> MetricValue:
-        nr_cards = self._metric_source.nr_of_over_due_actions()
+        nr_cards = self._metric_source.nr_of_over_due_actions(*self._get_metric_source_ids()) \
+            if self._metric_source else -1
         return -1 if nr_cards is None else nr_cards
 
     def url(self) -> Dict[str, str]:
-        return self._metric_source.over_due_actions_url() or {}
+        return self._metric_source.over_due_actions_url(*self._get_metric_source_ids()) \
+            if self._metric_source else {}
 
 
 class StaleActions(LowerIsBetterMetric):
@@ -94,9 +98,11 @@ class StaleActions(LowerIsBetterMetric):
     metric_source_class = metric_source.ActionLog
 
     def value(self) -> MetricValue:
-        nr_cards = self._metric_source.nr_of_inactive_actions()
+        nr_cards = self._metric_source.nr_of_inactive_actions(*self._get_metric_source_ids()) \
+            if self._metric_source else -1
         return -1 if nr_cards is None else nr_cards
 
     def url(self) -> Dict[str, str]:
-        return self._metric_source.inactive_actions_url() or {}
+        return self._metric_source.inactive_actions_url(*self._get_metric_source_ids()) \
+            if self._metric_source else {}